# shared/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "access_token",
          "refresh_token",
          "expires_at",
          "user"
        ],
        "title": "AuthSessionResponse",
        "type": "object"
      },
      "AuthUser": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_guest": {
            "title": "Is Guest",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "is_guest"
        ],
        "title": "AuthUser",
        "type": "object"
      },
      "DialogueResponse": {
        "additionalProperties": false,
        "properties": {
          "character_id": {
            "title": "Character Id",
            "type": "string"
          },
          "character_name": {
            "title": "Character Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "companion",
              "npc"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "character_id",
          "character_name",
          "role",
          "text"
        ],
        "title": "DialogueResponse",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GameIslandState": {
        "additionalProperties": false,
        "properties": {
          "island_id": {
            "const": "joy_meadow",
            "title": "Island Id",
            "type": "string"
          },
          "name": {
            "const": "Joy Meadow",
            "title": "Name",
            "type": "string"
          },
          "restoration_level": {
            "maximum": 100,
            "minimum": 0,
            "title": "Restoration Level",
            "type": "integer"
          },
          "restored": {
            "title": "Restored",
            "type": "boolean"
          }
        },
        "required": [
          "island_id",
          "name",
          "restoration_level",
          "restored"
        ],
        "title": "GameIslandState",
        "type": "object"
      },
      "GameStartRequest": {
        "additionalProperties": false,
        "properties": {
          "island_id": {
            "const": "joy_meadow",
            "default": "joy_meadow",
            "title": "Island Id",
            "type": "string"
          }
        },
        "title": "GameStartRequest",
        "type": "object"
      },
      "GameStateResponse": {
        "additionalProperties": false,
        "properties": {
          "dialogue": {
            "items": {
              "$ref": "#/components/schemas/DialogueResponse"
            },
            "title": "Dialogue",
            "type": "array"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/InventoryItemResponse"
            },
            "title": "Inventory",
            "type": "array"
          },
          "island": {
            "$ref": "#/components/schemas/GameIslandState"
          },
          "journal": {
            "items": {
              "$ref": "#/components/schemas/JournalEntryResponse"
            },
            "title": "Journal",
            "type": "array"
          },
          "quest": {
            "$ref": "#/components/schemas/QuestStateResponse"
          },
          "recipe": {
            "$ref": "#/components/schemas/RecipeStateResponse"
          },
          "save": {
            "$ref": "#/components/schemas/SaveStatusResponse"
          },
          "started": {
            "title": "Started",
            "type": "boolean"
          }
        },
        "required": [
          "started",
          "island",
          "dialogue",
          "inventory",
          "quest",
          "recipe",
          "journal",
          "save"
        ],
        "title": "GameStateResponse",
        "type": "object"
      },
      "GuestLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Display Name"
          }
        },
        "title": "GuestLoginRequest",
        "type": "object"
      },
...
      "InventoryCollectRequest": {
        "additionalProperties": false,
        "properties": {
          "ingredient_id": {
            "enum": [
              "vanilla_orchid",
              "honey_bloom"
            ],
            "title": "Ingredient Id",
            "type": "string"
          }
        },
        "required": [
          "ingredient_id"
        ],
        "title": "InventoryCollectRequest",
        "type": "object"
      },
      "InventoryItemResponse": {
        "additionalProperties": false,
        "properties": {
          "collected": {
            "title": "Collected",
            "type": "boolean"
          },
          "ingredient_id": {
            "enum": [
              "vanilla_orchid",
              "honey_bloom"
            ],
            "title": "Ingredient Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quantity": {
            "minimum": 0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "ingredient_id",
          "name",
          "quantity",
          "collected"
        ],
        "title": "InventoryItemResponse",
        "type": "object"
      },
      "InventoryResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InventoryItemResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "InventoryResponse",
        "type": "object"
      },
...
      "JournalEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "recipe_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Name"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "emotion",
          "recipe_name",
          "created_at"
        ],
        "title": "JournalEntryResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LogoutRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "minLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Refresh Token"
          }
        },
        "title": "LogoutRequest",
        "type": "object"
      },
      "LogoutResponse": {
        "additionalProperties": false,
        "properties": {
          "logged_out": {
            "title": "Logged Out",
            "type": "boolean"
          }
        },
        "required": [
          "logged_out"
        ],
        "title": "LogoutResponse",
        "type": "object"
      },
...
      "MeResponse": {
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        },
        "required": [
          "user"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "QuestStateResponse": {
        "additionalProperties": false,
        "properties": {
          "can_complete": {
            "title": "Can Complete",
            "type": "boolean"
          },
          "collected_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Collected Ingredients",
            "type": "array"
          },
          "crafted": {
            "title": "Crafted",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "quest_id": {
            "title": "Quest Id",
            "type": "string"
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "string"
          },
          "required_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Required Ingredients",
            "type": "array"
          },
          "status": {
            "enum": [
              "not_started",
              "active",
              "completed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "quest_id",
          "title",
          "description",
          "status",
          "required_ingredients",
          "collected_ingredients",
          "recipe_id",
          "crafted",
          "can_complete"
        ],
        "title": "QuestStateResponse",
        "type": "object"
      },
...
      "RecipeStateResponse": {
        "additionalProperties": false,
        "properties": {
          "ability": {
            "title": "Ability",
            "type": "string"
          },
          "can_craft": {
            "title": "Can Craft",
            "type": "boolean"
          },
          "crafted": {
            "title": "Crafted",
            "type": "boolean"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "lore": {
            "title": "Lore",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "string"
          },
          "required_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Required Ingredients",
            "type": "array"
          }
        },
        "required": [
          "recipe_id",
          "name",
          "emotion",
          "lore",
          "ability",
          "required_ingredients",
          "crafted",
          "can_craft"
        ],
        "title": "RecipeStateResponse",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "minLength": 32,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RefreshResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_at"
        ],
        "title": "RefreshResponse",
        "type": "object"
      },
      "RegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Flavor Odyssey Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/guest": {
      "post": {
        "operationId": "auth_guest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Guest Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "auth_logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LogoutRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "auth_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "auth_refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "auth_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/game/start": {
      "post": {
        "operationId": "start_game",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Game",
        "tags": [
          "game"
        ]
      }
    },
    "/api/game/state": {
      "get": {
        "operationId": "get_game_state",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get State",
        "tags": [
          "game"
        ]
      }
    },
    "/api/inventory": {
      "get": {
        "operationId": "get_inventory",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Inventory",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/inventory/collect": {
      "post": {
        "operationId": "collect_inventory_ingredient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryCollectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Collect Ingredient",
        "tags": [
          "inventory"
        ]
      }
    },
...
  },
...
}